// WADNR.Web/src/app/shared/generated/model/project-organization-item.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class ProjectOrganizationItem { 
    ProjectOrganizationID?: number;
    OrganizationID?: number;
    OrganizationName?: string | null;
    RelationshipTypeID?: number;
    RelationshipTypeName?: string | null;
    IsPrimaryContact?: boolean;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ProjectOrganizationItemForm { 
    ProjectOrganizationID?: FormControl<number>;
    OrganizationID?: FormControl<number>;
    OrganizationName?: FormControl<string>;
    RelationshipTypeID?: FormControl<number>;
    RelationshipTypeName?: FormControl<string>;
    IsPrimaryContact?: FormControl<boolean>;
}

export class ProjectOrganizationItemFormControls { 
    public static ProjectOrganizationID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static OrganizationID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static OrganizationName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static RelationshipTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static RelationshipTypeName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static IsPrimaryContact = (value: FormControlState<boolean> | boolean = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<boolean>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}