// WADNR.Web/src/app/shared/generated/model/project-organization-save-request.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ProjectOrganizationItemRequest } from './project-organization-item-request';


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class ProjectOrganizationSaveRequest { 
    Organizations?: Array<ProjectOrganizationItemRequest> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ProjectOrganizationSaveRequestForm { 
    Organizations?: FormControl<Array<ProjectOrganizationItemRequest>>;
}

export class ProjectOrganizationSaveRequestFormControls { 
    public static Organizations = (value: FormControlState<Array<ProjectOrganizationItemRequest>> | Array<ProjectOrganizationItemRequest> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<ProjectOrganizationItemRequest>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}