// WADNR.Web/src/app/shared/generated/model/project-organization-item-request.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class ProjectOrganizationItemRequest { 
    ProjectOrganizationID?: number | null;
    OrganizationID: number;
    RelationshipTypeID: number;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ProjectOrganizationItemRequestForm { 
    ProjectOrganizationID?: FormControl<number>;
    OrganizationID: FormControl<number>;
    RelationshipTypeID: FormControl<number>;
}

export class ProjectOrganizationItemRequestFormControls { 
    public static ProjectOrganizationID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static OrganizationID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: true,
            validators: 
            [
                Validators.required,
            ],
        }
    );
    public static RelationshipTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: true,
            validators: 
            [
                Validators.required,
            ],
        }
    );
}